utoipa

Module openapi

Rust implementation of Openapi Spec V3.1.

Re-exports

pub use self::content::Content;

pub use self::content::ContentBuilder;

pub use self::external_docs::ExternalDocs;

pub use self::header::Header;

pub use self::header::HeaderBuilder;

pub use self::info::Contact;

pub use self::info::ContactBuilder;

pub use self::info::Info;

pub use self::info::InfoBuilder;

pub use self::info::License;

pub use self::info::LicenseBuilder;

pub use self::path::HttpMethod;

pub use self::path::PathItem;

pub use self::path::Paths;

pub use self::path::PathsBuilder;

pub use self::response::Response;

pub use self::response::ResponseBuilder;

pub use self::response::Responses;

pub use self::response::ResponsesBuilder;

pub use self::schema::AllOf;

pub use self::schema::AllOfBuilder;

pub use self::schema::Array;

pub use self::schema::ArrayBuilder;

pub use self::schema::Components;

pub use self::schema::ComponentsBuilder;

pub use self::schema::Discriminator;

pub use self::schema::KnownFormat;

pub use self::schema::Object;

pub use self::schema::ObjectBuilder;

pub use self::schema::OneOf;

pub use self::schema::OneOfBuilder;

pub use self::schema::Ref;

pub use self::schema::Schema;

pub use self::schema::SchemaFormat;

pub use self::schema::ToArray;

pub use self::schema::Type;

pub use self::security::SecurityRequirement;

pub use self::server::Server;

pub use self::server::ServerBuilder;

pub use self::server::ServerVariable;

pub use self::server::ServerVariableBuilder;

pub use self::tag::Tag;

Modules

content

Implements content object for request body and response.

encoding

Implements encoding object for content.

example

Implements OpenAPI Example Object can be used to define examples for Responses and RequestBodys.

extensions

Implements OpenAPI Extensions.

external_docs

Implements OpenAPI External Docs Object types.

header

Implements OpenAPI Header Object types.

info

Implements OpenAPI Metadata types.

link

Implements Open API Link Object for responses.

path

Implements OpenAPI Path Object types.

request_body

Implements OpenAPI Request Body types.

response

Implements OpenApi Responses.

schema

Implements OpenAPI Schema Object types which can be used to define field properties, enum values, array or object types.

security

Implements OpenAPI Security Schema types.

server

Implements OpenAPI Server Object types to configure target servers.

tag

Implements OpenAPI Tag Object types.

xml

Implements OpenAPI Xml Object types.

Structs

OpenApi

Root object of the OpenAPI document.

OpenApiBuilder

Builder for OpenApi with chainable configuration methods to create a new OpenApi.

Enums

Deprecated

Value used to indicate whether reusable schema, parameter or operation is deprecated.

OpenApiVersion

Represents available OpenAPI versions.

RefOr

A Ref or some other type T.

Required

Value used to indicate whether parameter or property is required.