utoipa::openapi

Enum RefOr

pub enum RefOr<T> {
    Ref(Ref),
    T(T),
}

A Ref or some other type T.

Typically used in combination with Components and is an union type between Ref and any other given type such as Schema or Response.

Variants

Ref(Ref)

Represents Ref reference to another OpenAPI object instance. e.g. $ref: #/components/schemas/Hello

T(T)

Represents any value that can be added to the Components e.g. Schema or Response.