utoipa::openapi

Module example

Implements OpenAPI Example Object can be used to define examples for Responses and RequestBodys.

Structs

Example

Implements OpenAPI Example Object.

ExampleBuilder

Builder for Example with chainable configuration methods to create a new Example.