Struct utoipa::openapi::example::ExampleBuilder

pub struct ExampleBuilder { /* private fields */ }

Builder for Example with chainable configuration methods to create a new Example.

Examples

Construct a new Example via builder

let example = ExampleBuilder::new()
    .summary("Example string response")
    .value(Some(serde_json::json!("Example value")))
    .build();

Implementations

impl ExampleBuilder

pub fn new() -> ExampleBuilder

Constructs a new ExampleBuilder.

pub fn build(self) -> Example

Constructs a new Example taking all fields values from this object.

impl ExampleBuilder

pub fn summary<S: Into<String>>(self, summary: S) -> Self

Add or change a short description for the Example. Setting this to empty String will make it not render in the generated OpenAPI document.

pub fn description<D: Into<String>>(self, description: D) -> Self

Add or change a long description for the Example. Markdown syntax is supported for rich text representation.

Setting this to empty String will make it not render in the generated OpenAPI document.

pub fn value(self, value: Option<Value>) -> Self

Add or change embedded literal example value. Example::value and Example::external_value are mutually exclusive.

pub fn external_value<E: Into<String>>(self, external_value: E) -> Self

Add or change an URI that points to a literal example value. Example::external_value provides the capability to references an example that cannot be easily included in JSON or YAML. Example::value and Example::external_value are mutually exclusive.

Setting this to an empty String will make the field not to render in the generated OpenAPI document.