utoipa::openapi

Module info

Implements OpenAPI Metadata types.

Refer to OpenApi trait and derive documentation for examples and usage details.

Structs

Contact

OpenAPI Contact information of the API.

ContactBuilder

Builder for Contact with chainable configuration methods to create a new Contact. See the InfoBuilder for combined usage example.

Info

OpenAPI Info object represents metadata of the API.

InfoBuilder

Builder for Info with chainable configuration methods to create a new Info.

License

OpenAPI License information of the API.

LicenseBuilder

Builder for License with chainable configuration methods to create a new License.