Module utoipa::openapi::path

Implements OpenAPI Path Object types.

Structs

Operation

Implements OpenAPI Operation Object object.

OperationBuilder

Builder for Operation with chainable configuration methods to create a new Operation.

Parameter

Implements OpenAPI Parameter Object for Operation.

ParameterBuilder

Builder for Parameter with chainable configuration methods to create a new Parameter.

PathItem

Implements OpenAPI Path Item Object what describes Operations available on a single path.

PathItemBuilder

Builder for PathItem with chainable configuration methods to create a new PathItem.

Paths

Implements OpenAPI Paths Object.

PathsBuilder

Builder for Paths with chainable configuration methods to create a new Paths.

Enums

ParameterIn

In definition of Parameter.

ParameterStyle

Defines how Parameter should be serialized.

PathItemType

Path item operation type.