Struct utoipa::openapi::path::Operation

#[non_exhaustive]
pub struct Operation {
    pub tags: Option<Vec<String>>,
    pub summary: Option<String>,
    pub description: Option<String>,
    pub operation_id: Option<String>,
    pub external_docs: Option<ExternalDocs>,
    pub parameters: Option<Vec<Parameter>>,
    pub request_body: Option<RequestBody>,
    pub responses: Responses,
    pub callbacks: Option<String>,
    pub deprecated: Option<Deprecated>,
    pub security: Option<Vec<SecurityRequirement>>,
    pub servers: Option<Vec<Server>>,
}

Implements OpenAPI Operation Object object.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

tags: Option<Vec<String>>

List of tags used for grouping operations.

When used with derive #[utoipa::path(...)] attribute macro the default value used will be resolved from handler path provided in #[openapi(paths(...))] with #[derive(OpenApi)] macro. If path resolves to None value crate will be used by default.

summary: Option<String>

Short summary what Operation does.

When used with derive #[utoipa::path(...)] attribute macro the value is taken from first line of doc comment.

description: Option<String>

Long explanation of Operation behaviour. Markdown syntax is supported.

When used with derive #[utoipa::path(...)] attribute macro the doc comment is used as value for description.

operation_id: Option<String>

Unique identifier for the API Operation. Most typically this is mapped to handler function name.

When used with derive #[utoipa::path(...)] attribute macro the handler function name will be used by default.

external_docs: Option<ExternalDocs>

Additional external documentation for this operation.

parameters: Option<Vec<Parameter>>

List of applicable parameters for this Operation.

request_body: Option<RequestBody>

Optional request body for this Operation.

responses: Responses

List of possible responses returned by the Operation.

callbacks: Option<String>deprecated: Option<Deprecated>

Define whether the operation is deprecated or not and thus should be avoided consuming.

security: Option<Vec<SecurityRequirement>>

Declaration which security mechanisms can be used for for the operation. Only one SecurityRequirement must be met.

Security for the Operation can be set to optional by adding empty security with SecurityRequirement::default.

servers: Option<Vec<Server>>

Alternative Servers for this Operation.

Implementations

impl Operation

pub fn new() -> Self

Construct a new API Operation.