Struct utoipa::openapi::path::OperationBuilder

pub struct OperationBuilder { /* private fields */ }

Builder for Operation with chainable configuration methods to create a new Operation.

Implementations

impl OperationBuilder

pub fn new() -> OperationBuilder

Constructs a new OperationBuilder.

pub fn build(self) -> Operation

Constructs a new Operation taking all fields values from this object.

impl OperationBuilder

pub fn tags<I: IntoIterator<Item = String>>(self, tags: Option<I>) -> Self

Add or change tags of the Operation.

pub fn tag<S: Into<String>>(self, tag: S) -> Self

Append tag to Operation tags.

pub fn summary<S: Into<String>>(self, summary: Option<S>) -> Self

Add or change short summary of the Operation.

pub fn description<S: Into<String>>(self, description: Option<S>) -> Self

Add or change description of the Operation.

pub fn operation_id<S: Into<String>>(self, operation_id: Option<S>) -> Self

Add or change operation id of the Operation.

pub fn parameters<I: IntoIterator<Item = P>, P: Into<Parameter>>( self, parameters: Option<I> ) -> Self

Add or change parameters of the Operation.

pub fn parameter<P: Into<Parameter>>(self, parameter: P) -> Self

Append parameter to Operation parameters.

pub fn request_body(self, request_body: Option<RequestBody>) -> Self

Add or change request body of the Operation.

pub fn responses<R: Into<Responses>>(self, responses: R) -> Self

Add or change responses of the Operation.

pub fn response<S: Into<String>, R: Into<RefOr<Response>>>( self, code: S, response: R ) -> Self

Append status code and a Response to the Operation responses map.

code must be valid HTTP status code.

response is instances of Response.

pub fn deprecated(self, deprecated: Option<Deprecated>) -> Self

Add or change deprecated status of the Operation.

pub fn securities<I: IntoIterator<Item = SecurityRequirement>>( self, securities: Option<I> ) -> Self

Add or change list of SecurityRequirements that are available for Operation.

pub fn security(self, security: SecurityRequirement) -> Self

Append SecurityRequirement to Operation security requirements.

pub fn servers<I: IntoIterator<Item = Server>>(self, servers: Option<I>) -> Self

Add or change list of Servers of the Operation.

pub fn server(self, server: Server) -> Self

Append a new Server to the Operation servers.