utoipa::openapi::path

Struct PathItem

#[non_exhaustive]
pub struct PathItem {

    pub summary: Option<String>,
    pub description: Option<String>,
    pub servers: Option<Vec<Server>>,
    pub parameters: Option<Vec<Parameter>>,
    pub get: Option<Operation>,
    pub put: Option<Operation>,
    pub post: Option<Operation>,
    pub delete: Option<Operation>,
    pub options: Option<Operation>,
    pub head: Option<Operation>,
    pub patch: Option<Operation>,
    pub trace: Option<Operation>,
    pub extensions: Option<Extensions>,

}

Implements OpenAPI Path Item Object what describes Operations available on a single path.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

summary: Option<String>

Optional summary intended to apply all operations in this PathItem.

description: Option<String>

Optional description intended to apply all operations in this PathItem. Description supports markdown syntax.

servers: Option<Vec<Server>>

Alternative Server array to serve all Operations in this PathItem overriding the global server array.

parameters: Option<Vec<Parameter>>

List of Parameters common to all Operations in this PathItem. Parameters cannot contain duplicate parameters. They can be overridden in Operation level but cannot be removed there.

get: Option<Operation>

Get Operation for the PathItem.

put: Option<Operation>

Put Operation for the PathItem.

post: Option<Operation>

Post Operation for the PathItem.

delete: Option<Operation>

Delete Operation for the PathItem.

options: Option<Operation>

Options Operation for the PathItem.

head: Option<Operation>

Head Operation for the PathItem.

patch: Option<Operation>

Patch Operation for the PathItem.

trace: Option<Operation>

Trace Operation for the PathItem.

extensions: Option<Extensions>

Optional extensions โx-somethingโ.

Implementations

impl PathItem

pub fn builder() -> PathItemBuilder

Construct a new PathItemBuilder.

This is effectively same as calling PathItemBuilder::new

impl PathItem

pub fn new<O: Into<Operation>>(http_method: HttpMethod, operation: O) -> Self

Construct a new PathItem with provided Operation mapped to given HttpMethod.

pub fn from_http_methods<I: IntoIterator<Item = HttpMethod>, O: Into<Operation>>( http_methods: I, operation: O, ) -> Self

Constructs a new PathItem with given Operation set for provided HttpMethods.

pub fn merge_operations(&mut self, path_item: PathItem)

Merge all defined Operations from given PathItem to self if self does not have existing operation.