utoipa::openapi::path

Struct Paths

#[non_exhaustive]
pub struct Paths {
    pub paths: PathsMap<String, PathItem>,
    pub extensions: Option<Extensions>,
}

Implements OpenAPI Paths Object.

Holds relative paths to matching endpoints and operations. The path is appended to the url from Server object to construct a full url for endpoint.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

paths: PathsMap<String, PathItem>

Map of relative paths with PathItems holding Operations matching api endpoints.

extensions: Option<Extensions>

Optional extensions โx-somethingโ.

Implementations

impl Paths

pub fn builder() -> PathsBuilder

Construct a new PathsBuilder.

This is effectively same as calling PathsBuilder::new

impl Paths

pub fn new() -> Self

Construct a new Paths object.

pub fn get_path_item<P: AsRef<str>>(&self, path: P) -> Option<&PathItem>

Return Option of reference to PathItem by given relative path P if one exists in Paths::paths map. Otherwise will return None.

Examples

Get user path item.

let path_item = paths.get_path_item("/api/v1/user");

pub fn get_path_operation<P: AsRef<str>>( &self, path: P, http_method: HttpMethod, ) -> Option<&Operation>

Return Option of reference to Operation from map of paths or None if not found.

First will try to find PathItem by given relative path P e.g. "/api/v1/user".

Then tries to find Operation from PathItemโs operations by given HttpMethod.

Examples

Get user operation from paths.

let operation = paths.get_path_operation("/api/v1/user", HttpMethod::Get);

pub fn add_path_operation<P: AsRef<str>, O: Into<Operation>>( &mut self, path: P, http_methods: Vec<HttpMethod>, operation: O, )

Append path operation to the list of paths.

Method accepts three arguments; path to add operation for, http_methods list of allowed HTTP methods for the Operation and operation to be added under the path.

If path already exists, the provided Operation will be set to existing path item for given list of HttpMethods.

pub fn merge(&mut self, other_paths: Paths)

Merge other_paths into self. On conflicting path the path item operations will be merged into existing PathItem. Otherwise path with PathItem will be appended to self. All Extensions will be merged from other_paths into self.