utoipa::openapi

Module schema

Implements OpenAPI Schema Object types which can be used to define field properties, enum values, array or object types.

Structs

AllOf

AllOf Composite Object component holds multiple components together where API endpoint will return a combination of all of them.

AllOfBuilder

Builder for AllOf with chainable configuration methods to create a new AllOf.

AnyOf

AnyOf Composite Object component holds multiple components together where API endpoint will return a combination of one or more of them.

AnyOfBuilder

Builder for AnyOf with chainable configuration methods to create a new AnyOf.

Array

Array represents Vec or slice type of items.

ArrayBuilder

Builder for Array with chainable configuration methods to create a new Array.

Components

Implements OpenAPI Components Object which holds supported reusable objects.

ComponentsBuilder

Builder for Components with chainable configuration methods to create a new Components.

Discriminator

OpenAPI Discriminator object which can be optionally used together with OneOf composite object.

Object

Implements subset of OpenAPI Schema Object which allows adding other Schemas as properties to this Schema.

ObjectBuilder

Builder for Object with chainable configuration methods to create a new Object.

OneOf

OneOf Composite Object component holds multiple components together where API endpoint could return any of them.

OneOfBuilder

Builder for OneOf with chainable configuration methods to create a new OneOf.

Ref

Implements OpenAPI Reference Object that can be used to reference reusable components such as Schemas or Responses.

RefBuilder

Builder for Ref with chainable configuration methods to create a new Ref.

Enums

AdditionalProperties

AdditionalProperties is used to define values of map fields of the Schema.

ArrayItems

Represents Array items in JSON Schema Array.

KnownFormat

Known schema format modifier property to provide fine detail of the primitive type.

Schema

Is super type for OpenAPI Schema Object. Schema is reusable resource what can be referenced from path operations and other components using Ref.

SchemaFormat

Additional format for SchemaType to fine tune the data type used. If the format is not supported by the UI it may default back to SchemaType alone. Format is an open value, so you can use any formats, even not those defined by the OpenAPI Specification.

SchemaType

Represents type of Schema.

Type

Represents data type fragment of Schema.

Traits

ToArray

This convenience trait allows quick way to wrap any RefOr<Schema> with Array schema.

Functions

empty

Create an empty Schema that serializes to null.