utoipa::openapi::schema

Enum KnownFormat

pub enum KnownFormat {

    Int32,
    Int64,
    Float,
    Double,
    Byte,
    Binary,
    Time,
    Date,
    DateTime,
    Duration,
    Password,
    Uuid,
    Email,
    IdnEmail,
    Hostname,
    IdnHostname,
    Ipv4,
    Ipv6,
    UriTemplate,
    JsonPointer,
    RelativeJsonPointer,
    Regex,

}

Known schema format modifier property to provide fine detail of the primitive type.

Known format is defined in https://spec.openapis.org/oas/latest.html#data-types and https://datatracker.ietf.org/doc/html/draft-bhutton-json-schema-validation-00#section-7.3 as well as by few known data types that are enabled by specific feature flag e.g. uuid.

Variants

Int32

32 bit integer.

Int64

64 bit integer.

Float

floating point number.

Double

double (floating point) number.

Byte

base64 encoded chars.

Binary

binary data (octet).

Time

ISO-8601 full time format RFC3339.

Date

ISO-8601 full date RFC3339.

DateTime

ISO-8601 full date time RFC3339.

Duration

duration format from RFC3339 Appendix-A.

Password

Hint to UI to obscure input.

Uuid

Used with String values to indicate value is in UUID format.

uuid feature need to be enabled.

Email

As defined in โMailboxโ rule RFC5321.

IdnEmail

As defined by extended โMailboxโ rule RFC6531.

Hostname

As defined by RFC1123, including host names produced using the Punycode algorithm specified in RFC5891.

IdnHostname

As defined by either RFC1123 as for hostname, or an internationalized hostname as defined by RFC5890.

Ipv4

An IPv4 address according to RFC2673.

Ipv6

An IPv6 address according to RFC4291.

UriTemplate

A string instance is a valid URI Template if it is according to RFC6570.

Note! There are no separate IRL template.

JsonPointer

A valid JSON string representation of a JSON Pointer according to RFC6901.

RelativeJsonPointer

A valid relative JSON Pointer according to draft-handrews-relative-json-pointer-01.

Regex

Regular expression, which SHOULD be valid according to the ECMA-262.