utoipa::openapi::schema

Struct AllOfBuilder

pub struct AllOfBuilder { /* private fields */ }

Builder for AllOf with chainable configuration methods to create a new AllOf.

Implementations

impl AllOfBuilder

pub fn new() -> AllOfBuilder

Constructs a new AllOfBuilder.

pub fn build(self) -> AllOf

Constructs a new AllOf taking all fields values from this object.

impl AllOfBuilder

pub fn item<I: Into<RefOr<Schema>>>(self, component: I) -> Self

Adds a given Schema to AllOf Composite Object.

pub fn schema_type<T: Into<SchemaType>>(self, schema_type: T) -> Self

Add or change type of the object e.g. to change type to string use value SchemaType::Type(Type::String).

pub fn title<I: Into<String>>(self, title: Option<I>) -> Self

Add or change the title of the AllOf.

pub fn description<I: Into<String>>(self, description: Option<I>) -> Self

Add or change optional description for AllOf component.

pub fn default(self, default: Option<Value>) -> Self

Add or change default value for the object which is provided when user has not provided the input in Swagger UI.

pub fn example(self, example: Option<Value>) -> Self

๐Deprecated: Since OpenAPI 3.1 prefer using examples

Add or change example shown in UI of the value for richer documentation.

Deprecated since 3.0.x. Prefer AllOfBuilder::examples instead

pub fn examples<I: IntoIterator<Item = V>, V: Into<Value>>( self, examples: I, ) -> Self

Add or change examples shown in UI of the value for richer documentation.

pub fn discriminator(self, discriminator: Option<Discriminator>) -> Self

Add or change discriminator field of the composite AllOf type.

pub fn extensions(self, extensions: Option<Extensions>) -> Self

Add openapi extensions (x-something) for AllOf.

pub fn to_array_builder(self) -> ArrayBuilder

Construct a new ArrayBuilder with this component set to ArrayBuilder::items.