utoipa::openapi::schema

Struct ArrayBuilder

pub struct ArrayBuilder { /* private fields */ }

Builder for Array with chainable configuration methods to create a new Array.

Implementations

impl ArrayBuilder

pub fn new() -> ArrayBuilder

Constructs a new ArrayBuilder.

pub fn build(self) -> Array

Constructs a new Array taking all fields values from this object.

impl ArrayBuilder

pub fn items<I: Into<ArrayItems>>(self, items: I) -> Self

Set Schema type for the Array.

pub fn prefix_items<I: IntoIterator<Item = S>, S: Into<Schema>>( self, items: I, ) -> Self

Add prefix items of Array to define item validation of tuples according JSON schema item validation.

pub fn schema_type<T: Into<SchemaType>>(self, schema_type: T) -> Self

Change type of the array e.g. to change type to string use value SchemaType::Type(Type::String).

Examples

Make nullable string array.

let _ = ArrayBuilder::new()
    .schema_type(SchemaType::from_iter([Type::Array, Type::Null]))
    .items(Object::with_type(Type::String))
    .build();

pub fn title<I: Into<String>>(self, title: Option<I>) -> Self

Add or change the title of the Array.

pub fn description<I: Into<String>>(self, description: Option<I>) -> Self

Add or change description of the property. Markdown syntax is supported.

pub fn deprecated(self, deprecated: Option<Deprecated>) -> Self

Add or change deprecated status for Array.

pub fn example(self, example: Option<Value>) -> Self

๐Deprecated: Since OpenAPI 3.1 prefer using examples

Add or change example shown in UI of the value for richer documentation.

Deprecated since 3.0.x. Prefer Array::examples instead

pub fn examples<I: IntoIterator<Item = V>, V: Into<Value>>( self, examples: I, ) -> Self

Add or change examples shown in UI of the value for richer documentation.

pub fn default(self, default: Option<Value>) -> Self

Add or change default value for the object which is provided when user has not provided the input in Swagger UI.

pub fn max_items(self, max_items: Option<usize>) -> Self

Set maximum allowed length for Array.

pub fn min_items(self, min_items: Option<usize>) -> Self

Set minimum allowed length for Array.

pub fn unique_items(self, unique_items: bool) -> Self

Set or change whether Array should enforce all items to be unique.

pub fn xml(self, xml: Option<Xml>) -> Self

Set Xml formatting for Array.

pub fn content_encoding<S: Into<String>>(self, content_encoding: S) -> Self

Set of change Object::content_encoding. Typically left empty but could be base64 for example.

pub fn content_media_type<S: Into<String>>(self, content_media_type: S) -> Self

Set of change Object::content_media_type. Value must be valid MIME type e.g. application/json.

pub fn extensions(self, extensions: Option<Extensions>) -> Self

Add openapi extensions (x-something) for Array.

pub fn to_array_builder(self) -> ArrayBuilder

Construct a new ArrayBuilder with this component set to ArrayBuilder::items.