utoipa::openapi::schema

Struct ObjectBuilder

pub struct ObjectBuilder { /* private fields */ }

Builder for Object with chainable configuration methods to create a new Object.

Implementations

impl ObjectBuilder

pub fn new() -> ObjectBuilder

Constructs a new ObjectBuilder.

pub fn build(self) -> Object

Constructs a new Object taking all fields values from this object.

impl ObjectBuilder

pub fn schema_type<T: Into<SchemaType>>(self, schema_type: T) -> Self

Add or change type of the object e.g. to change type to string use value SchemaType::Type(Type::String).

pub fn format(self, format: Option<SchemaFormat>) -> Self

Add or change additional format for detailing the schema type.

pub fn property<S: Into<String>, I: Into<RefOr<Schema>>>( self, property_name: S, component: I, ) -> Self

Add new property to the Object.

Method accepts property name and property component as an arguments.

pub fn additional_properties<I: Into<AdditionalProperties<Schema>>>( self, additional_properties: Option<I>, ) -> Self

Add additional Schema for non specified fields (Useful for typed maps).

pub fn property_names<S: Into<Schema>>(self, property_name: Option<S>) -> Self

Add additional Schema to describe property names of an object such as a map. See more details https://json-schema.org/draft/2020-12/draft-bhutton-json-schema-01#name-propertynames

pub fn required<I: Into<String>>(self, required_field: I) -> Self

Add field to the required fields of Object.

pub fn title<I: Into<String>>(self, title: Option<I>) -> Self

Add or change the title of the Object.

pub fn description<I: Into<String>>(self, description: Option<I>) -> Self

Add or change description of the property. Markdown syntax is supported.

pub fn default(self, default: Option<Value>) -> Self

Add or change default value for the object which is provided when user has not provided the input in Swagger UI.

pub fn deprecated(self, deprecated: Option<Deprecated>) -> Self

Add or change deprecated status for Object.

pub fn enum_values<I: IntoIterator<Item = E>, E: Into<Value>>( self, enum_values: Option<I>, ) -> Self

Add or change enum property variants.

pub fn example(self, example: Option<Value>) -> Self

๐Deprecated: Since OpenAPI 3.1 prefer using examples

Add or change example shown in UI of the value for richer documentation.

Deprecated since 3.0.x. Prefer Object::examples instead

pub fn examples<I: IntoIterator<Item = V>, V: Into<Value>>( self, examples: I, ) -> Self

Add or change examples shown in UI of the value for richer documentation.

pub fn write_only(self, write_only: Option<bool>) -> Self

Add or change write only flag for Object.

pub fn read_only(self, read_only: Option<bool>) -> Self

Add or change read only flag for Object.

pub fn xml(self, xml: Option<Xml>) -> Self

Add or change additional Xml formatting of the Object.

pub fn multiple_of<N: Into<Number>>(self, multiple_of: Option<N>) -> Self

Set or change multiple_of validation flag for number and integer type values.

pub fn maximum<N: Into<Number>>(self, maximum: Option<N>) -> Self

Set or change inclusive maximum value for number and integer values.

pub fn minimum<N: Into<Number>>(self, minimum: Option<N>) -> Self

Set or change inclusive minimum value for number and integer values.

pub fn exclusive_maximum<N: Into<Number>>( self, exclusive_maximum: Option<N>, ) -> Self

Set or change exclusive maximum value for number and integer values.

pub fn exclusive_minimum<N: Into<Number>>( self, exclusive_minimum: Option<N>, ) -> Self

Set or change exclusive minimum value for number and integer values.

pub fn max_length(self, max_length: Option<usize>) -> Self

Set or change maximum length for string values.

pub fn min_length(self, min_length: Option<usize>) -> Self

Set or change minimum length for string values.

pub fn pattern<I: Into<String>>(self, pattern: Option<I>) -> Self

Set or change a valid regular expression for string value to match.

pub fn max_properties(self, max_properties: Option<usize>) -> Self

Set or change maximum number of properties the Object can hold.

pub fn min_properties(self, min_properties: Option<usize>) -> Self

Set or change minimum number of properties the Object can hold.

pub fn extensions(self, extensions: Option<Extensions>) -> Self

Add openapi extensions (x-something) for Object.

pub fn content_encoding<S: Into<String>>(self, content_encoding: S) -> Self

Set of change Object::content_encoding. Typically left empty but could be base64 for example.

pub fn content_media_type<S: Into<String>>(self, content_media_type: S) -> Self

Set of change Object::content_media_type. Value must be valid MIME type e.g. application/json.

pub fn to_array_builder(self) -> ArrayBuilder

Construct a new ArrayBuilder with this component set to ArrayBuilder::items.