utoipa::openapi::schema

Struct Ref

#[non_exhaustive]
pub struct Ref {
    pub ref_location: String,
    pub description: String,
    pub summary: String,
}

Implements OpenAPI Reference Object that can be used to reference reusable components such as Schemas or Responses.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

ref_location: String

Reference location of the actual component.

description: String

A description which by default should override that of the referenced component. Description supports markdown syntax. If referenced object type does not support description this field does not have effect.

summary: String

A short summary which by default should override that of the referenced component. If referenced component does not support summary field this does not have effect.

Implementations

impl Ref

pub fn builder() -> RefBuilder

Construct a new RefBuilder.

This is effectively same as calling RefBuilder::new

impl Ref

pub fn new<I: Into<String>>(ref_location: I) -> Self

Construct a new Ref with custom ref location. In most cases this is not necessary and Ref::from_schema_name could be used instead.

pub fn from_schema_name<I: Into<String>>(schema_name: I) -> Self

Construct a new Ref from provided schema name. This will create a Ref that references the the reusable schemas.

pub fn from_response_name<I: Into<String>>(response_name: I) -> Self

Construct a new Ref from provided response name. This will create a Ref that references the reusable response.

pub fn to_array_builder(self) -> ArrayBuilder

Construct a new ArrayBuilder with this component set to ArrayBuilder::items.