Module utoipa::openapi::security

Implements OpenAPI Security Schema types.

Refer to SecurityScheme for usage and more details.

Structs

ApiKeyValue

Value object for ApiKey.

AuthorizationCode

Authorization code Flow configuration for OAuth2.

ClientCredentials

Client credentials Flow configuration for OAuth2.

Http

Http authentication SecurityScheme builder.

HttpBuilder

Builder for Http with chainable configuration methods to create a new Http.

Implicit

Implicit Flow configuration for OAuth2.

OAuth2

OAuth2 Flow configuration for SecurityScheme.

OpenIdConnect

Open id connect SecurityScheme

Password

Password Flow configuration for OAuth2.

Scopes

OAuth2 flow scopes object defines required permissions for oauth flow.

SecurityRequirement

OpenAPI security requirement object.

Enums

ApiKey

Api key authentication SecurityScheme.

Flow

OAuth2 flow configuration object.

HttpAuthScheme

Implements types according RFC7235.

SecurityScheme

OpenAPI security scheme for path operations.