utoipa::openapi::server

Struct Server

#[non_exhaustive]
pub struct Server {
    pub url: String,
    pub description: Option<String>,
    pub variables: Option<BTreeMap<String, ServerVariable>>,
    pub extensions: Option<Extensions>,
}

Represents target server object. It can be used to alter server connection for path operations.

By default OpenAPI will implicitly implement Server with url = "/" if no servers is provided to the OpenApi.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

url: String

Target url of the Server. It can be valid http url or relative path.

Url also supports variable substitution with {variable} syntax. The substitutions then can be configured with Server::variables map.

description: Option<String>

Optional description describing the target server url. Description supports markdown syntax.

variables: Option<BTreeMap<String, ServerVariable>>

Optional map of variable name and its substitution value used in Server::url.

extensions: Option<Extensions>

Optional extensions โx-somethingโ.

Implementations

impl Server

pub fn builder() -> ServerBuilder

Construct a new ServerBuilder.

This is effectively same as calling ServerBuilder::new

impl Server

pub fn new<S: Into<String>>(url: S) -> Self

Construct a new Server with given url. Url can be valid http url or context path of the url.

If url is valid http url then all path operation requestโs will be forwarded to the selected Server.

If url is path of url e.g. /api/v1 then the url will be appended to the servers address and the operations will be forwarded to location server address + url.

Examples

Create new server with url path.

 Server::new("/api/v1");

Create new server with alternative server.

 Server::new("https://alternative.pet-api.test/api/v1");