Struct utoipa::openapi::server::ServerBuilder

pub struct ServerBuilder { /* private fields */ }

Builder for Server with chainable configuration methods to create a new Server.

Implementations

impl ServerBuilder

pub fn new() -> ServerBuilder

Constructs a new ServerBuilder.

pub fn build(self) -> Server

Constructs a new Server taking all fields values from this object.

impl ServerBuilder

pub fn url<U: Into<String>>(self, url: U) -> Self

Add url to the target Server.

pub fn description<S: Into<String>>(self, description: Option<S>) -> Self

Add or change description of the Server.

pub fn parameter<N: Into<String>, V: Into<ServerVariable>>( self, name: N, variable: V ) -> Self

Add parameter to Server which is used to substitute values in Server::url.

name Defines name of the parameter which is being substituted within the url. If url has {username} substitution then the name should be username.

parameter Use ServerVariableBuilder to define how the parameter is being substituted within the url.