utoipa::openapi::server

Struct ServerVariableBuilder

pub struct ServerVariableBuilder { /* private fields */ }

Builder for ServerVariable with chainable configuration methods to create a new ServerVariable.

Implementations

impl ServerVariableBuilder

pub fn new() -> ServerVariableBuilder

Constructs a new ServerVariableBuilder.

pub fn build(self) -> ServerVariable

Constructs a new ServerVariable taking all fields values from this object.

impl ServerVariableBuilder

pub fn default_value<S: Into<String>>(self, default_value: S) -> Self

Add default value for substitution.

pub fn description<S: Into<String>>(self, description: Option<S>) -> Self

Add or change description of substituted parameter.

pub fn enum_values<I: IntoIterator<Item = V>, V: Into<String>>( self, enum_values: Option<I>, ) -> Self

Add or change possible values used to substitute parameter.

pub fn extensions(self, extensions: Option<Extensions>) -> Self

Add openapi extensions (x-something) of the API.