utoipa::openapi::xml

Struct Xml

#[non_exhaustive]
pub struct Xml {
    pub name: Option<Cow<'static, str>>,
    pub namespace: Option<Cow<'static, str>>,
    pub prefix: Option<Cow<'static, str>>,
    pub attribute: Option<bool>,
    pub wrapped: Option<bool>,
}

Implements OpenAPI Xml Object.

Can be used to modify xml output format of specific OpenAPI Schema Object which are implemented in schema module.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Option<Cow<'static, str>>

Used to replace the name of attribute or type used in schema property. When used with Xml::wrapped attribute the name will be used as a wrapper name for wrapped array instead of the item or type name.

namespace: Option<Cow<'static, str>>

Valid uri definition of namespace used in xml.

prefix: Option<Cow<'static, str>>

Prefix for xml element Xml::name.

attribute: Option<bool>

Flag deciding will this attribute translate to element attribute instead of xml element.

wrapped: Option<bool>

Flag only usable with array definition. If set to true the output xml will wrap the array of items <pets><pet></pet></pets> instead of unwrapped <pet></pet>.

Implementations

impl Xml

pub fn builder() -> XmlBuilder

Construct a new XmlBuilder.

This is effectively same as calling XmlBuilder::new

impl Xml

pub fn new() -> Self

Construct a new Xml object.