utoipa

Trait IntoResponses

pub trait IntoResponses {
    // Required method
    fn responses() -> BTreeMap<String, RefOr<Response>>;
}

This trait is implemented to document a type (like an enum) which can represent multiple responses, to be used in operation.

Examples

use std::collections::BTreeMap;
use utoipa::{
    openapi::{Response, ResponseBuilder, ResponsesBuilder, RefOr},
    IntoResponses,
};

enum MyResponse {
    Ok,
    NotFound,
}

impl IntoResponses for MyResponse {
    fn responses() -> BTreeMap<String, RefOr<Response>> {
        ResponsesBuilder::new()
            .response("200", ResponseBuilder::new().description("Ok"))
            .response("404", ResponseBuilder::new().description("Not Found"))
            .build()
            .into()
    }
}

Required Methods

fn responses() -> BTreeMap<String, RefOr<Response>>

Returns an ordered map of response codes to responses.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.