Struct utoipa::openapi::content::ContentBuilder

pub struct ContentBuilder { /* private fields */ }

Builder for Content with chainable configuration methods to create a new Content.

Implementations

impl ContentBuilder

pub fn new() -> ContentBuilder

Constructs a new ContentBuilder.

pub fn build(self) -> Content

Constructs a new Content taking all fields values from this object.

impl ContentBuilder

pub fn schema<I: Into<RefOr<Schema>>>(self, component: I) -> Self

Add schema.

pub fn example(self, example: Option<Value>) -> Self

Add example of schema.

pub fn examples_from_iter<E: IntoIterator<Item = (N, V)>, N: Into<String>, V: Into<RefOr<Example>>>( self, examples: E ) -> Self

Add iterator of (N, V) where N is name of example and V is Example to Content of a request body or response body.

Content::examples and Content::example are mutually exclusive. If both are defined examples will override value in example.

pub fn encoding<S: Into<String>, E: Into<Encoding>>( self, property_name: S, encoding: E ) -> Self

Add an encoding.

The property_name MUST exist in the Content::schema as a property, with schema being a Schema::Object and this object containing the same property key in Object::properties.

The encoding object SHALL only apply to request_body objects when the media type is multipart or application/x-www-form-urlencoded.