Struct utoipa::openapi::path::ParameterBuilder

pub struct ParameterBuilder { /* private fields */ }

Builder for Parameter with chainable configuration methods to create a new Parameter.

Implementations

impl ParameterBuilder

pub fn new() -> ParameterBuilder

Constructs a new ParameterBuilder.

pub fn build(self) -> Parameter

Constructs a new Parameter taking all fields values from this object.

impl ParameterBuilder

pub fn name<I: Into<String>>(self, name: I) -> Self

Add name of the Parameter.

pub fn parameter_in(self, parameter_in: ParameterIn) -> Self

Add in of the Parameter.

pub fn required(self, required: Required) -> Self

Add required declaration of the Parameter. If ParameterIn::Path is defined this is always Required::True.

pub fn description<S: Into<String>>(self, description: Option<S>) -> Self

Add or change description of the Parameter.

pub fn deprecated(self, deprecated: Option<Deprecated>) -> Self

Add or change Parameter deprecated declaration.

pub fn schema<I: Into<RefOr<Schema>>>(self, component: Option<I>) -> Self

Add or change Parameters schema.

pub fn style(self, style: Option<ParameterStyle>) -> Self

Add or change serialization style of Parameter.

pub fn explode(self, explode: Option<bool>) -> Self

Define whether Parameters are exploded or not.

pub fn allow_reserved(self, allow_reserved: Option<bool>) -> Self

Add or change whether Parameter should allow reserved characters.

pub fn example(self, example: Option<Value>) -> Self

Add or change example of Parameter’s potential value.