Struct utoipa::openapi::schema::ArrayBuilder

pub struct ArrayBuilder { /* private fields */ }

Builder for Array with chainable configuration methods to create a new Array.

Implementations

impl ArrayBuilder

pub fn new() -> ArrayBuilder

Constructs a new ArrayBuilder.

pub fn build(self) -> Array

Constructs a new Array taking all fields values from this object.

impl ArrayBuilder

pub fn items<I: Into<RefOr<Schema>>>(self, component: I) -> Self

Set Schema type for the Array.

pub fn title<I: Into<String>>(self, title: Option<I>) -> Self

Add or change the title of the Array.

pub fn description<I: Into<String>>(self, description: Option<I>) -> Self

Add or change description of the property. Markdown syntax is supported.

pub fn deprecated(self, deprecated: Option<Deprecated>) -> Self

Add or change deprecated status for Array.

pub fn example(self, example: Option<Value>) -> Self

Add or change example shown in UI of the value for richer documentation.

pub fn default(self, default: Option<Value>) -> Self

Add or change default value for the object which is provided when user has not provided the input in Swagger UI.

pub fn max_items(self, max_items: Option<usize>) -> Self

Set maximum allowed length for Array.

pub fn min_items(self, min_items: Option<usize>) -> Self

Set minimum allowed length for Array.

pub fn unique_items(self, unique_items: bool) -> Self

Set or change whether Array should enforce all items to be unique.

pub fn xml(self, xml: Option<Xml>) -> Self

Set Xml formatting for Array.

pub fn nullable(self, nullable: bool) -> Self

Add or change nullable flag for Object.

pub fn to_array_builder(self) -> ArrayBuilder

Construct a new ArrayBuilder with this component set to ArrayBuilder::items.